use keydown to close settings
I don't know why I keep noticing this but a key up should never close something like this. Because when I select text if I keyup outside the modal I close the whole thing when I was just trying to select my text, it just happened so I ended up here. The action to close the modal is to click on the overlay... so why not use the event that works on click?